fix(frontend): inline dashboard-v4 types into hook, remove restored directory
dashboard-v4 components were deleted in PR #64. Import should not
have restored the deleted directory — types now inlined in the
only consumer (use-dashboard-v4.ts).
