--- frontend/components/dashboard-v4/index.ts
@@ -1,6 +1,0 @@
-/**
- * FILE: frontend/components/dashboard-v4/index.ts
- * PURPOSE: Re-export Dashboard V4 types and components
- */
-
-export * from "./types";
